# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/feedbacks:
    get:
      tags:
        - Feedbacks
      summary: List Feedbacks
      description: Get a paginated list of Feedbacks with optional filtering and sorting
      operationId: listFeedbacks
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Feedbacks
      summary: Create Feedback
      description: Create a new Feedback
      operationId: createFeedback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFeedbackRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Feedback'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/goals:
    get:
      tags:
        - Goals
      summary: List Goals
      description: Get a paginated list of Goals with optional filtering and sorting
      operationId: listGoals
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoalList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Goals
      summary: Create Goal
      description: Create a new Goal
      operationId: createGoal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGoalRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Goal'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    Feedback:
      type: object
      required:
        - id
        - from_employee_id
        - to_employee_id
        - content
        - is_anonymous
        - relationship
        - metadata
      properties:
        id:
          type: string
          format: uuid
        cycle_id:
          nullable: true
          type: string
          format: uuid
        from_employee_id:
          type: string
          format: uuid
        to_employee_id:
          type: string
          format: uuid
        content:
          type: string
        is_anonymous:
          type: boolean
        relationship:
          type: string
        metadata:
          type: object
...
    CreateFeedbackRequest:
      type: object
      required:
        - from_employee_id
        - to_employee_id
        - content
        - is_anonymous
        - relationship
        - metadata
      properties:
        cycle_id:
          nullable: true
          type: string
          format: uuid
        from_employee_id:
          type: string
          format: uuid
        to_employee_id:
          type: string
          format: uuid
        content:
          type: string
        is_anonymous:
          type: boolean
        relationship:
          type: string
        metadata:
          type: object
...
    FeedbackList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Feedback'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    Goal:
      type: object
      required:
        - id
        - employee_id
        - title
        - status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        employee_id:
          type: string
          format: uuid
        cycle_id:
          nullable: true
          type: string
          format: uuid
        title:
          type: string
          maxLength: 200
        description:
          nullable: true
          type: string
        weight:
          nullable: true
          type: string
        progress:
          nullable: true
          type: string
        parent_goal_id:
          nullable: true
          type: string
          format: uuid
        status:
          type: string
        metadata:
          type: object
...
    CreateGoalRequest:
      type: object
      required:
        - employee_id
        - title
        - status
        - metadata
      properties:
        employee_id:
          type: string
          format: uuid
        cycle_id:
          nullable: true
          type: string
          format: uuid
        title:
          type: string
          maxLength: 200
        description:
          nullable: true
          type: string
        weight:
          nullable: true
          type: string
        progress:
          nullable: true
          type: string
        parent_goal_id:
          nullable: true
          type: string
          format: uuid
        status:
          type: string
        metadata:
          type: object
...
    GoalList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Goal'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []